rename merge_run_config.py -> clustering_run_config.py
