+++ spd/clustering/scripts/run_clustering.py
@@ -31,6 +31,7 @@ from spd.clustering.activations import (
     component_activations,
     process_activations,
 )
+from spd.clustering.clustering_run_config import ClusteringRunConfig
 from spd.clustering.consts import (
     ActivationsTensor,
     BatchTensor,
@@ -43,7 +44,6 @@ from spd.clustering.math.merge_matrix import GroupMerge
 from spd.clustering.math.semilog import semilog
 from spd.clustering.merge import merge_iteration
 from spd.clustering.merge_history import MergeHistory
-from spd.clustering.merge_run_config import ClusteringRunConfig
 from spd.clustering.plotting.activations import plot_activations
 from spd.clustering.plotting.merge import plot_merge_history_cluster_sizes, plot_merge_iteration
 from spd.clustering.storage import StorageBase
